Relocate test utilities.
