+++ benchmarks/tests/tpcds_correctness_test.rs
@@ -5,10 +5,10 @@ mod tests {
     use datafusion::error::Result;
     use datafusion::physical_plan::{ExecutionPlan, collect};
     use datafusion::prelude::SessionContext;
-    use datafusion_distributed::test_utils::in_memory_channel_resolver::start_in_memory_context;
     use datafusion_distributed::test_utils::property_based::{
         compare_ordering, compare_result_set,
     };
+    use datafusion_distributed::test_utils::start_in_memory_context;
     use datafusion_distributed::{
         DefaultSessionBuilder, DistributedExec, DistributedExt, display_plan_ascii,
     };
