+++ benchmarks/tests/tpch_plans_test.rs
@@ -1,6 +1,6 @@
 #[cfg(all(feature = "tpch", test))]
 mod tests {
-    use datafusion_distributed::test_utils::in_memory_channel_resolver::start_in_memory_context;
+    use datafusion_distributed::test_utils::start_in_memory_context;
     use datafusion_distributed::{
         DefaultSessionBuilder, DistributedExt, assert_snapshot, display_plan_ascii,
     };
