+++ examples/custom_distributed_partial_reduction_tree.rs
@@ -52,9 +52,8 @@ use datafusion::physical_plan::aggregates::{AggregateExec, AggregateMode, Physic
 use datafusion::physical_plan::coalesce_partitions::CoalescePartitionsExec;
 use datafusion::physical_plan::expressions::Column;
 use datafusion::prelude::{ParquetReadOptions, SessionContext};
-use datafusion_distributed::test_utils::in_memory_channel_resolver::{
-    InMemoryChannelResolver, InMemoryWorkerResolver,
-};
+use datafusion_distributed::test_utils::InMemoryChannelResolver;
+use datafusion_distributed::test_utils::in_memory_channel_resolver::InMemoryWorkerResolver;
 use datafusion_distributed::{
     DistributedExt, NetworkCoalesceExec, SessionStateBuilderExt, WorkerQueryContext,
     display_plan_ascii,
