+++ src/execution_plans/benchmarks/shuffle_bench.rs
@@ -2,9 +2,9 @@ use super::fixture::{
     InMemoryChannelsResolver, benchmark_schema, make_input_partitions, rows_for_producer,
 };
 use crate::common::task_ctx_with_extension;
+use crate::protocol::grpc::test_utils::worker_handles::MemoryWorkerHandle;
 use crate::stage::RemoteStage;
 use crate::worker::WorkerConnectionPool;
-use crate::worker::test_utils::worker_handles::MemoryWorkerHandle;
 use crate::{DistributedExt, DistributedTaskContext, NetworkShuffleExec, Stage};
 use arrow::datatypes::Schema;
 use arrow_ipc::CompressionType;
