+++ src/protocol/grpc/mod.rs
@@ -5,6 +5,8 @@ mod metrics_proto;
 mod observability;
 mod on_drop_stream;
 mod spawn_select_all;
+#[cfg(any(test, feature = "integration"))]
+pub mod test_utils;
 mod worker_client;
 mod worker_service;
