+++ src/protocol/grpc/test_utils/mod.rs
@@ -1,0 +1,3 @@
+pub mod in_memory_channel_resolver;
+pub mod localhost;
+pub mod worker_handles;
