+++ src/worker/task_data.rs
@@ -23,11 +23,11 @@ pub struct TaskData {
     /// Sender half of the metrics channel. `impl_coordinator_channel` takes this (via
     /// `Option::take`) when the coordinator channel reaches EOS, sending the collected metrics
     /// back to the coordinator through the `CoordinatorChannel` side channel.
-    pub(super) metrics_tx: Arc<std::sync::Mutex<Option<oneshot::Sender<TaskMetrics>>>>,
+    pub(crate) metrics_tx: Arc<std::sync::Mutex<Option<oneshot::Sender<TaskMetrics>>>>,
     /// Metrics related to the execution of a task within a stage. This metrics, instead of being
     /// associated to a specific node, they are global to the task, like the time at which the plan
     /// was fed by the coordinator to the worker.
-    pub(super) task_data_metrics: Arc<TaskDataMetrics>,
+    pub(crate) task_data_metrics: Arc<TaskDataMetrics>,
 }
 
 pub(crate) const PLAN_ADDED_AT_METRIC: &str = "plan_added_at";
@@ -35,7 +35,7 @@ pub(crate) const PLAN_EXECUTED_AT_METRIC: &str = "plan_executed_at";
 pub(crate) const PLAN_FINISHED_AT_METRIC: &str = "plan_finished_at";
 
 #[derive(Debug)]
-pub(super) struct TaskDataMetrics {
+pub(crate) struct TaskDataMetrics {
     pub(super) query_start_time_ns: usize,
     /// When the plan was set by the coordinator.
     pub(super) plan_added_at: MaxLatencyMetric,
@@ -46,7 +46,7 @@ pub(super) struct TaskDataMetrics {
 }
 
 impl TaskDataMetrics {
-    pub(super) fn new(query_start_time_ns: usize) -> Self {
+    pub(crate) fn new(query_start_time_ns: usize) -> Self {
         let plan_added_at = MaxLatencyMetric::default();
         plan_added_at.add_duration(Duration::from_nanos(
             now_ns::<u64>().saturating_sub(query_start_time_ns as u64),
