--- src/worker/test_utils/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod worker_handles;
